requirements.txt
aioredis==2.0.1
amqp==5.1.1
async-timeout==4.0.2
bidict==0.22.0
billiard==3.6.4.0
cachelib==0.9.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.2.1
eventlet==0.33.2
Flask==2.2.2
Flask-Session==0.4.0
Flask-SocketIO==5.3.2
flask-talisman==1.0.0
Flask-WTF==1.0.1
greenlet==2.0.1
hiredis==2.1.0
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
more-itertools==8.14.0
pptree==3.1
prompt-toolkit==3.0.36
pydantic==1.10.2
pygame==2.1.2
python-decouple==3.7
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
python-ulid==1.1.0
pytz==2022.7
redis==4.4.0
redis-om==0.1.1
six==1.16.0
types-redis==4.3.21.6
typing_extensions==4.4.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
whitenoise==6.3.0
WTForms==3.0.1
zipp==3.11.0
